Port Gsuite Rules to ActivityEvent (#326)
