+++ gsuite_activityevent_rules/gsuite_user_suspended.py
@@ -1,6 +1,4 @@
 from panther_base_helpers import deep_get
-from panther_base_helpers import gsuite_details_lookup as details_lookup
-from panther_base_helpers import gsuite_parameter_lookup as param_lookup
 
 USER_SUSPENDED_EVENTS = {
     "account_disabled_generic",
@@ -14,12 +12,11 @@ def rule(event):
     if deep_get(event, "id", "applicationName") != "login":
         return False
 
-    return bool(details_lookup("account_warning", USER_SUSPENDED_EVENTS, event))
+    return bool(event.get("name") in USER_SUSPENDED_EVENTS)
 
 
 def title(event):
-    details = details_lookup("account_warning", USER_SUSPENDED_EVENTS, event)
-    user = param_lookup(details.get("parameters", {}), "affected_email_address")
+    user = deep_get(event, "parameters", "affected_email_address")
     if not user:
         user = "<UNKNOWN_USER>"
     return f"User [{user}]'s account was disabled"
